requirements.txt
aiogram>=3.4,<4.0
curl_cffi>=0.7,<1.0
beautifulsoup4>=4.12,<5.0
python-dotenv>=1.0,<2.0
